Submission Job | FrontEnd

+++ database/migrations/2023_06_26_093302_create_jobs_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->text('detail');
             $table->foreignId('category_id');
             $table->foreignId('location_id');
-            $table->unsignedInteger('rate');
+            $table->unsignedBigInteger('rate');
             $table->string('phone');
             $table->string('option_one')->nullable();
             $table->string('option_two')->nullable();
